Fully remove accessories integration
