+++ src/main/java/com/wenxin2/marioverse/client/renderers/curios/OneUpRenderer.java
@@ -1,8 +1,7 @@
-package com.wenxin2.marioverse.client.renderers.accesories;
+package com.wenxin2.marioverse.client.renderers.curios;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wenxin2.marioverse.registries.ConfigRegistry;
-import io.wispforest.accessories.api.slot.SlotReference;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.HumanoidModel;
@@ -19,29 +18,6 @@ import top.theillusivec4.curios.api.SlotContext;
 import top.theillusivec4.curios.api.client.ICurioRenderer;
 
 public class OneUpRenderer implements ICurioRenderer, SimpleItemRenderer {
-    @Override
-    public <M extends LivingEntity> void render(ItemStack stack, SlotReference slotReference, PoseStack poseStack,
-                                                EntityModel<M> model, MultiBufferSource buffer,
-                                                int light, float limbSwing, float limbSwingAmount, float partialTicks,
-                                                float ageInTicks, float netHeadYaw, float headPitch) {
-        if (ConfigRegistry.RENDER_ONE_UP_CHARM.get()) {
-            poseStack.pushPose();
-                this.translateIfSneaking(poseStack, slotReference.entity());
-                EntityRenderer<? super LivingEntity> render = Minecraft.getInstance().getEntityRenderDispatcher().getRenderer(slotReference.entity());
-                if (render instanceof LivingEntityRenderer) {
-                    if (model instanceof HumanoidModel humanoidModel)
-                        this.rotateIfSneaking(poseStack, slotReference.entity(), humanoidModel);
-                }
-                poseStack.mulPose(Direction.DOWN.getRotation());
-                poseStack.translate(0.15F, -0.45F, 0.145F);
-                poseStack.scale(0.25F, 0.25F, 0.45F);
-                Minecraft.getInstance().getItemRenderer()
-                        .renderStatic(stack, ItemDisplayContext.NONE, light, OverlayTexture.NO_OVERLAY,
-                                poseStack, buffer, slotReference.entity().level(), 0);
-            poseStack.popPose();
-        }
-    }
-
     @Override
     public <T extends LivingEntity, M extends EntityModel<T>> void render(ItemStack stack, SlotContext slotContext,
            PoseStack poseStack, RenderLayerParent<T, M> renderLayerParent, MultiBufferSource buffer,
