+++ src/main/java/com/wenxin2/marioverse/curios/PlasticBucketCurio.java
@@ -1,0 +1,12 @@
+package com.wenxin2.marioverse.curios;
+
+import net.minecraft.world.item.ItemStack;
+import top.theillusivec4.curios.api.SlotContext;
+import top.theillusivec4.curios.api.type.capability.ICurioItem;
+
+public class PlasticBucketCurio implements ICurioItem {
+    @Override
+    public boolean canEquipFromUse(SlotContext slotContext, ItemStack stack) {
+        return false;
+    }
+}
