+++ src/main/java/com/wenxin2/marioverse/datagen/EntityTypeTagsGen.java
@@ -417,7 +417,6 @@ public class EntityTypeTagsGen extends EntityTypeTagsProvider {
                 .add(EntityType.ZOMBIE)
                 .add(EntityType.ZOMBIE_VILLAGER)
                 .add(EntityType.ZOMBIFIED_PIGLIN)
-                .addOptionalTag(CompatRegistry.ACCESSORIES_DEFAULT_TARGETS)
                 .addOptionalTag(CompatRegistry.TWILIGHT_FOREST_BOSSES)
                 .addOptional(GUARD_VILLAGER)
                 .addOptional(FEMALE_VILLAGER)
@@ -454,7 +453,6 @@ public class EntityTypeTagsGen extends EntityTypeTagsProvider {
                 .add(EntityType.ZOMBIE)
                 .add(EntityType.ZOMBIE_VILLAGER)
                 .add(EntityType.ZOMBIFIED_PIGLIN)
-                .addOptionalTag(CompatRegistry.ACCESSORIES_DEFAULT_TARGETS)
                 .addOptionalTag(CompatRegistry.TWILIGHT_FOREST_BOSSES)
                 .addOptional(GUARD_VILLAGER)
                 .addOptional(FEMALE_VILLAGER)
@@ -484,7 +482,6 @@ public class EntityTypeTagsGen extends EntityTypeTagsProvider {
                 .add(EntityType.ZOMBIE)
                 .add(EntityType.ZOMBIE_VILLAGER)
                 .add(EntityType.ZOMBIFIED_PIGLIN)
-                .addOptionalTag(CompatRegistry.ACCESSORIES_DEFAULT_TARGETS)
                 .addOptionalTag(CompatRegistry.TWILIGHT_FOREST_BOSSES)
                 .addOptional(GUARD_VILLAGER)
                 .addOptional(FEMALE_VILLAGER)
@@ -513,7 +510,6 @@ public class EntityTypeTagsGen extends EntityTypeTagsProvider {
                 .add(EntityType.ZOMBIE)
                 .add(EntityType.ZOMBIE_VILLAGER)
                 .add(EntityType.ZOMBIFIED_PIGLIN)
-                .addOptionalTag(CompatRegistry.ACCESSORIES_DEFAULT_TARGETS)
                 .addOptionalTag(CompatRegistry.TWILIGHT_FOREST_BOSSES)
                 .addOptional(GUARD_VILLAGER)
                 .addOptional(FEMALE_VILLAGER)
@@ -542,7 +538,6 @@ public class EntityTypeTagsGen extends EntityTypeTagsProvider {
                 .add(EntityType.ZOMBIE)
                 .add(EntityType.ZOMBIE_VILLAGER)
                 .add(EntityType.ZOMBIFIED_PIGLIN)
-                .addOptionalTag(CompatRegistry.ACCESSORIES_DEFAULT_TARGETS)
                 .addOptionalTag(CompatRegistry.TWILIGHT_FOREST_BOSSES)
                 .addOptional(GUARD_VILLAGER)
                 .addOptional(FEMALE_VILLAGER)
