+++ src/main/java/com/wenxin2/marioverse/integration/CompatRegistry.java
@@ -59,7 +59,6 @@ public class CompatRegistry {
     public static final Supplier<SoundEvent> CONFETTI_POPPER_SOUND = make("supplementaries:item.confetti_popper", BuiltInRegistries.SOUND_EVENT);
     public static final Supplier<SoundEvent> ICE_BOMB_SOUND = make("twilightforest:item.twilightforest.ice_bomb.fired", BuiltInRegistries.SOUND_EVENT);
 
-    public static final TagKey<EntityType<?>> ACCESSORIES_DEFAULT_TARGETS = entityTypeTags("accessories", "defaulted_targets");
     public static final TagKey<EntityType<?>> TWILIGHT_FOREST_BOSSES = entityTypeTags("twilightforest", "bosses");
 
     public static final TagKey<Block> CREATE_BRITTLE = TagRegistry.blockTags("create", "brittle");
