+++ src/main/java/com/wenxin2/marioverse/mixin/HumanoidArmorLayerMixin.java
@@ -3,28 +3,16 @@ package com.wenxin2.marioverse.mixin;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.blaze3d.vertex.VertexConsumer;
-import com.wenxin2.marioverse.client.renderers.SuperStarRenderType;
-import com.wenxin2.marioverse.client.renderers.accesories.ArmorRenderingExtension;
-import io.wispforest.accessories.AccessoriesLoaderInternals;
-import io.wispforest.accessories.compat.GeckoLibCompat;
+import com.wenxin2.marioverse.client.renderers.curios.ArmorRenderingExtension;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.client.renderer.entity.RenderLayerParent;
 import net.minecraft.client.renderer.entity.layers.HumanoidArmorLayer;
 import net.minecraft.client.renderer.entity.layers.RenderLayer;
 import net.minecraft.client.renderer.texture.OverlayTexture;
-import net.minecraft.core.component.DataComponents;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.ArmorItem;
-import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.component.DyedItemColor;
-import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -83,12 +71,5 @@ public abstract class HumanoidArmorLayerMixin<T extends LivingEntity, M extends 
             at = { @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;getItemBySlot(Lnet/minecraft/world/entity/EquipmentSlot;)Lnet/minecraft/world/item/ItemStack;")})
     private ItemStack getAlternativeStack(LivingEntity instance, EquipmentSlot equipmentSlot, Operation<ItemStack> original) {
         return this.tempStack != null ? this.tempStack : original.call(instance, equipmentSlot);
-    }
-
-    @Unique
-    private boolean attemptGeckoRender(ItemStack stack, PoseStack poseStack, MultiBufferSource multiBufferSource, T livingEntity, EquipmentSlot equipmentSlot, int light, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch) {
-        return GeckoLibCompat.renderGeckoArmor(poseStack, multiBufferSource, livingEntity, stack, equipmentSlot,
-                (HumanoidModel) this.getParentModel(), this.getArmorModel(equipmentSlot), partialTicks, light, limbSwing,
-                limbSwingAmount, ageInTicks, netHeadYaw, headPitch, this::setPartVisibility);
     }
 }
