+++ src/main/java/com/wenxin2/marioverse/registries/TagRegistry.java
@@ -151,20 +151,15 @@ public class TagRegistry {
     public static final TagKey<Item> CAN_SELECT_CLEAR_WARP_PIPES = itemTags("can_select_clear_warp_pipes");
     public static final TagKey<Item> CAN_SELECT_WATER_SPOUTS = itemTags("can_select_water_spouts");
     public static final TagKey<Item> CHARACTER_TRIMS = itemTags("character_trims");
-    public static final TagKey<Item> CHARMS = itemTags("accessories", "charm");
     public static final TagKey<Item> CHARMS_CURIO = itemTags("curios", "charm");
     public static final TagKey<Item> CHECKPOINT_FLAG_ITEMS = itemTags("checkpoint_flags");
     public static final TagKey<Item> CHEEP_CHEEP_ITEMS = itemTags("cheep_cheeps");
     public static final TagKey<Item> CHISELED_CALCITE_BRICK_ITEMS = itemTags("chiseled_calcite_bricks");
     public static final TagKey<Item> CORAL_TOWERS = itemTags("coral_towers");
     public static final TagKey<Item> COSTUMES = itemTags("costumes");
-    public static final TagKey<Item> COSTUME_HAT = itemTags("accessories", "costume_hat");
     public static final TagKey<Item> COSTUME_HAT_CURIO = itemTags("curios", "costume_hat");
-    public static final TagKey<Item> COSTUME_PANTS = itemTags("accessories", "costume_pants");
     public static final TagKey<Item> COSTUME_PANTS_CURIO = itemTags("curios", "costume_pants");
-    public static final TagKey<Item> COSTUME_SHIRT = itemTags("accessories", "costume_shirt");
     public static final TagKey<Item> COSTUME_SHIRT_CURIO = itemTags("curios", "costume_shirt");
-    public static final TagKey<Item> COSTUME_SHOES = itemTags("accessories", "costume_shoes");
     public static final TagKey<Item> COSTUME_SHOES_CURIO = itemTags("curios", "costume_shoes");
     public static final TagKey<Item> CRACKED_CALCITE_BRICK_ITEMS = itemTags("cracked_calcite_bricks");
     public static final TagKey<Item> CRAFTS_WARP_DOOR = itemTags("crafts_warp_door");
